data-generator/requirements.txt
flask
flask_cors
mysql-connector-python==8.0.15
numpy